requirements.txt
aiogram>=3.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
qrcode>=7.0
Pillow>=10.0
